SHARED(Frontend): OPHOTRKEH-242 Table update and refactor
