+++ frontend/packages/shared/src/components/index.tsx
@@ -29,8 +29,9 @@ export { OPHLogoViewer } from './OPHLogoViewer/OPHLogoViewer';
 export { DropDownMenuButton } from './DropDownMenuButton/DropDownMenuButton';
 export { SkipLink } from './SkipLink/SkipLink';
 export { Svg } from './Svg/Svg';
-export { PaginatedTable, NormalTable } from './Table/Table';
-export { ManagedPaginatedTable } from './Table/ManagedTable';
+export { CustomTable } from './Table/CustomTable';
+export { PaginatedTable } from './Table/PaginatedTable';
+export { ManagedPaginatedTable } from './Table/ManagedPaginatedTable';
 export { ToggleFilterGroup } from './ToggleFilterGroup/ToggleFilterGroup';
 export { LangSelector } from './LangSelector/LangSelector';
 export { InfoText } from './InfoText/InfoText';
